+++ src/components/Main/index.js
@@ -1,0 +1,5 @@
+import ContextMenu from './ContextMenu/ContextMenu';
+import MovieDetails from './MovieDetails/MovieDetails';
+import MovieTile from './MovieTile/MovieTile';
+import SortControl from './SortControl/SortControl';
+export { ContextMenu, MovieDetails, MovieTile, SortControl };
